+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/MessageConverter.kt
@@ -2,6 +2,7 @@ package jp.shiguredo.sora.sdk.channel.signaling.message
 
 import com.google.gson.Gson
 import jp.shiguredo.sora.sdk.channel.option.SoraChannelRole
+import jp.shiguredo.sora.sdk.channel.option.SoraForwardingFilterOption
 import jp.shiguredo.sora.sdk.channel.option.SoraMediaOption
 import jp.shiguredo.sora.sdk.error.SoraDisconnectReason
 import jp.shiguredo.sora.sdk.util.SoraLogger
@@ -39,7 +40,8 @@ class MessageConverter {
             bundleId: String? = null,
             signalingNotifyMetadata: Any? = null,
             dataChannels: List<Map<String, Any>>? = null,
-            redirect: Boolean = false
+            redirect: Boolean = false,
+            forwardingFilterOption: SoraForwardingFilterOption? = null,
         ): String {
 
             val msg = ConnectMessage(
@@ -55,6 +57,7 @@ class MessageConverter {
                 bundleId = bundleId,
                 signalingNotifyMetadata = signalingNotifyMetadata,
                 audioStreamingLanguageCode = mediaOption.audioStreamingLanguageCode,
+                forwardingFilter = forwardingFilterOption?.signaling,
             )
 
             if (mediaOption.upstreamIsRequired) {
@@ -74,6 +77,9 @@ class MessageConverter {
                 if (mediaOption.videoUpstreamEnabled) {
                     val videoSetting = VideoSetting(mediaOption.videoCodec.toString())
                     mediaOption.videoBitrate?.let { videoSetting.bitRate = it }
+                    mediaOption.videoVp9Params?.let { videoSetting.vp9Params = it }
+                    mediaOption.videoAv1Params?.let { videoSetting.av1Params = it }
+                    mediaOption.videoH264Params?.let { videoSetting.h264Params = it }
                     msg.video = videoSetting
                 } else {
                     msg.video = false